Made the `HistogramCache` use the generic cache interface as well
