+++ Jantimizer/Histograms/Managers/EquiDepthVarianceHistogramManager.cs
@@ -1,11 +1,11 @@
 ﻿using Histograms.DataGatherers;
 using Histograms.Models;
-using Histograms.Services;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using Tools.Caches;
 using Tools.Models;
 
 namespace Histograms.Managers
@@ -24,7 +24,7 @@ namespace Histograms.Managers
                 IDepthHistogram newHistogram = new HistogramEquiDepthVariance(tableName, attributeName, Depth);
                 newHistogram.GenerateHistogramFromSortedGroups(await DataGatherer.GetSortedGroupsFromDb(tableName, attributeName));
                 string columnHash = await DataGatherer.GetTableAttributeColumnHash(tableName, attributeName);
-                HistogramCache.AddToCacheIfNotThere(tableName, attributeName, columnHash, newHistogram);
+                HistogramCache.Instance.AddToCacheIfNotThere(new string[] { tableName, attributeName, columnHash }, newHistogram);
                 AddHistogram(newHistogram);
             }
             else
